Move `PrivateCredential` to the `credential` module
